pyproject.toml
[project]
name = "llm-rag-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.14.8",
    "python-dotenv>=1.2.1",
    "llama-index-readers-file>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-embeddings-huggingface>=0.3.1",
    "arxiv>=2.1.0",
]
